Fixed cyclic deps making turbo hang, fixed actions package cyclic deps, and fixed fast-average-color-node importing on the client.
