+++ apps/dapp/util/index.ts
@@ -1,2 +1,1 @@
 export * from './addToken'
-export * from './colors'
