--- packages/actions/actions/ActionKey.tsx
@@ -1,12 +1,0 @@
-export enum ActionKey {
-  Spend = 'spend',
-  Mint = 'mint',
-  Stake = 'stake',
-  AddToken = 'addToken',
-  RemoveToken = 'removeToken',
-  UpdateInfo = 'updateInfo',
-  UpdateProposalConfig = 'updateProposalConfig',
-  Instantiate = 'instantiate',
-  Execute = 'execute',
-  Custom = 'custom',
-}
