+++ packages/actions/hooks/useActionsForVotingModuleType.ts
@@ -2,7 +2,8 @@ import { useMemo } from 'react'
 
 import { VotingModuleType } from '@dao-dao/utils'
 
-import { Action, actions } from '..'
+import { actions } from '../actions'
+import { Action } from '../types'
 
 export const useActionsForVotingModuleType = (
   type: VotingModuleType
