+++ packages/ui/package.json
@@ -11,8 +11,6 @@
     "@dao-dao/icons": "*",
     "@dao-dao/types": "^0.0.8",
     "@dao-dao/utils": "*",
-    "@dao-dao/state": "*",
-    "@dao-dao/actions": "*",
     "@heroicons/react": "^1.0.5",
     "clsx": "^1.1.1",
     "codemirror": "^5.65.0",
